added new module "cluster"
