+++ tests/integration_test.rs
@@ -9,7 +9,8 @@ fn cluster_insertion_test() {
 }
 #[test]
 fn cluster_connection_test(){
+    use dxtracker::cluster::{connect};
     let call = "DD5HT";
     let cluster = "cluster.dl9gtb.de:8000";
-    assert_eq!(dxtracker::connect_to_cluster(cluster, call).is_ok(), true);
+    assert_eq!(connect(cluster, call).is_ok(), true);
 }
